% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{describe_attack}
\alias{describe_attack}
\title{Describes the details of a DDoS attack}
\usage{
describe_attack(AttackId)
}
\arguments{
\item{AttackId}{[required] The unique identifier (ID) for the attack that to be described.}
}
\description{
Describes the details of a DDoS attack.
}
\section{Accepted Parameters}{
\preformatted{describe_attack(
  AttackId = "string"
)
}
}

