% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{describe_emergency_contact_settings}
\alias{describe_emergency_contact_settings}
\title{Lists the email addresses that the DRT can use to contact you during a suspected attack}
\usage{
describe_emergency_contact_settings()
}
\description{
Lists the email addresses that the DRT can use to contact you during a suspected attack.
}
\section{Accepted Parameters}{
\preformatted{describe_emergency_contact_settings()
}
}

