% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{describe_protection}
\alias{describe_protection}
\title{Lists the details of a Protection object}
\usage{
describe_protection(ProtectionId = NULL, ResourceArn = NULL)
}
\arguments{
\item{ProtectionId}{The unique identifier (ID) for the Protection object that is described. When submitting the \code{DescribeProtection} request you must provide either the \code{ResourceArn} or the \code{ProtectionID}, but not both.}

\item{ResourceArn}{The ARN (Amazon Resource Name) of the AWS resource for the Protection object that is described. When submitting the \code{DescribeProtection} request you must provide either the \code{ResourceArn} or the \code{ProtectionID}, but not both.}
}
\description{
Lists the details of a Protection object.
}
\section{Accepted Parameters}{
\preformatted{describe_protection(
  ProtectionId = "string",
  ResourceArn = "string"
)
}
}

