% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{describe_subscription}
\alias{describe_subscription}
\title{Provides details about the AWS Shield Advanced subscription for an account}
\usage{
describe_subscription()
}
\description{
Provides details about the AWS Shield Advanced subscription for an account.
}
\examples{

}
