% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{get_subscription_state}
\alias{get_subscription_state}
\title{Returns the \code{SubscriptionState}, either \code{Active} or \code{Inactive}}
\usage{
get_subscription_state()
}
\description{
Returns the \code{SubscriptionState}, either \code{Active} or \code{Inactive}.
}
\examples{

}
