% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{list_attacks}
\alias{list_attacks}
\title{Returns all ongoing DDoS attacks or all DDoS attacks during a specified time period}
\usage{
list_attacks(ResourceArns = NULL, StartTime = NULL, EndTime = NULL,
  NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{ResourceArns}{The ARN (Amazon Resource Name) of the resource that was attacked. If this is left blank, all applicable resources for this account will be included.}

\item{StartTime}{The start of the time period for the attacks. This is a \code{timestamp} type. The sample request above indicates a \code{number} type because the default used by WAF is Unix time in seconds. However any valid \href{http://docs.aws.amazon.com/cli/latest/userguide/cli-using-param.html#parameter-types}{timestamp format} is allowed.}

\item{EndTime}{The end of the time period for the attacks. This is a \code{timestamp} type. The sample request above indicates a \code{number} type because the default used by WAF is Unix time in seconds. However any valid \href{http://docs.aws.amazon.com/cli/latest/userguide/cli-using-param.html#parameter-types}{timestamp format} is allowed.}

\item{NextToken}{The \code{ListAttacksRequest.NextMarker} value from a previous call to \code{ListAttacksRequest}. Pass null if this is the first call.}

\item{MaxResults}{The maximum number of AttackSummary objects to be returned. If this is left blank, the first 20 results will be returned.}
}
\description{
Returns all ongoing DDoS attacks or all DDoS attacks during a specified time period.
}
\examples{

}
