% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{list_protections}
\alias{list_protections}
\title{Lists all Protection objects for the account}
\usage{
list_protections(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The \code{ListProtectionsRequest.NextToken} value from a previous call to \code{ListProtections}. Pass null if this is the first call.}

\item{MaxResults}{The maximum number of Protection objects to be returned. If this is left blank the first 20 results will be returned.}
}
\description{
Lists all Protection objects for the account.
}
\examples{

}
