% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_package.R
\docType{package}
\name{paws.shield-package}
\alias{paws.shield}
\alias{paws.shield-package}
\title{paws.shield: AWS Shield}
\description{
This is the *AWS Shield Advanced API Reference*. This guide
is for developers who need detailed information about the AWS Shield
Advanced API actions, data types, and errors. For detailed information
about AWS WAF and AWS Shield Advanced features and an overview of how
to use the AWS WAF and AWS Shield Advanced APIs, see the [AWS WAF and
AWS Shield Developer
Guide](http://docs.aws.amazon.com/waf/latest/developerguide/).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
