% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{update_emergency_contact_settings}
\alias{update_emergency_contact_settings}
\title{Updates the details of the list of email addresses that the DRT can use to contact you during a suspected attack}
\usage{
update_emergency_contact_settings(EmergencyContactList = NULL)
}
\arguments{
\item{EmergencyContactList}{A list of email addresses that the DRT can use to contact you during a suspected attack.}
}
\description{
Updates the details of the list of email addresses that the DRT can use to contact you during a suspected attack.
}
\examples{

}
