% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{delete_protection}
\alias{delete_protection}
\title{Deletes an AWS Shield Advanced Protection}
\usage{
delete_protection(ProtectionId)
}
\arguments{
\item{ProtectionId}{[required] The unique identifier (ID) for the Protection object to be deleted.}
}
\description{
Deletes an AWS Shield Advanced Protection.
}
\section{Accepted Parameters}{
\preformatted{delete_protection(
  ProtectionId = "string"
)
}
}

