% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{describe_protection}
\alias{describe_protection}
\title{Lists the details of a Protection object}
\usage{
describe_protection(ProtectionId)
}
\arguments{
\item{ProtectionId}{The unique identifier (ID) for the Protection object that is described.}
}
\description{
Lists the details of a Protection object.
}
\examples{

}
