% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.shield_operations.R
\name{get_subscription_state}
\alias{get_subscription_state}
\title{Returns the SubscriptionState, either Active or Inactive}
\usage{
get_subscription_state()
}
\description{
Returns the \code{SubscriptionState}, either \code{Active} or \code{Inactive}.
}
\section{Accepted Parameters}{
\preformatted{get_subscription_state()
}
}

